XK.KA module API

Warning

This documentation is work in progress. Please refer to examples to get started.

XK.KA.BLOCK_SIZE

XK.KA.JL_CU_CONTEXT_SIZE

XK.KA.VERSION_STRUCT_FN_OFFSET

XK.KA.VERSION_STRUCT_MOODULE_OFFSET

XK.KA.FormatStruct

XK.KA.VersionStruct

XK.KA.Format

XK.KA.deinit

XK.KA.device_async

XK.KA.eval

XK.KA.get_arguments_tuple_size

XK.KA.get_returned_tuple_size

XK.KA.include

XK.KA.init

XK.KA.task_ka_launcher

XK.KA.task_ka_main

XK.KA.@kernel

XK.KA.@tid

XK.KA — Module

No documentation provided.

XK.KA.BLOCK_SIZE — Constant

No documentation provided.

XK.KA.JL_CU_CONTEXT_SIZE — Constant

No documentation provided.

XK.KA.VERSION_STRUCT_FN_OFFSET — Constant

No documentation provided.

XK.KA.VERSION_STRUCT_MOODULE_OFFSET — Constant

No documentation provided.

XK.KA.FormatStruct — Type

No documentation provided.

XK.KA.VersionStruct — Type

No documentation provided.

XK.KA.Format — Function

No documentation provided.

XK.KA.deinit — Function

No documentation provided.

XK.KA.device_async — Function

No documentation provided.

XK.KA.eval — Function

No documentation provided.

XK.KA.get_arguments_tuple_size — Function

No documentation provided.

XK.KA.get_returned_tuple_size — Function

No documentation provided.

XK.KA.include — Function

No documentation provided.

XK.KA.init — Function

No documentation provided.

XK.KA.task_ka_launcher — Function

No documentation provided.

XK.KA.task_ka_main — Function

No documentation provided.

XK.KA.@kernel — Macro

@KA.kernel function my_kernel(...) ... end A macro that wraps a function definition, adds 'return nothing', and works by manually inspecting the Abstract Syntax Tree (AST).

XK.KA.@tid — Macro

@KA.tid A macro that expands to the backend-specific index calculation.